When searching a species in the searchbar, update the genus too
This is a long outstanding bug that Eleanor found. Basically, when
selecting a species from the search input dropdown, I was only updating
the selected species, not it's parent genus and family too. Here's the
bug report:

"When looking at a genus and the species within that genus, the genus
that i’m looking at is bolded in the search list (as well as the species
that I select to look at). However when I go to search a new species in
the query, the genus and species in the search list doesn’t update to
bold that new genus and species.

For Instance:
Im looking at "Ficus fig" (genus) and more specifically the species
"common fig". Both these specifications are bolded in the search list.

Now I want to see the species coast redwood, so I type into the search
bar “coast redwood” and click the suggestion.

The page updates to show me coast redwood, but the search list is still
bolded on "Ficus fig”… Thus I cannot easily browse all of the other
species under the same genus as the coast redwood.

It would be nice if the search list updated to bold the genus and
species for coast redwood when I search for it."